+++ libNOM.map/Mapping.cs
@@ -1,5 +1,6 @@
 ﻿using libNOM.map.Extensions;
 using libNOM.map.Json;
+using libNOM.map.Services;
 using Newtonsoft.Json.Linq;
 
 namespace libNOM.map;
@@ -12,7 +13,7 @@ public static class Mapping
 {
     #region Field
 
-    private static readonly HttpClient _httpClient = new();
+    private static GitHubService? _gitHub;
     private static readonly MappingJson _jsonCompiler = MappingJson.Deserialize(Properties.Resources.MBINCompiler)!; // latest MBINCompiler mapping.json when this version was created
     private static MappingJson? _jsonDownload; // dynamic content from the latest MBINCompiler release on GitHub
     private static readonly MappingJson _jsonLegacy = MappingJson.Deserialize(Properties.Resources.Legacy)!; // older keys that are not present in the latest version
@@ -29,6 +30,8 @@ public static class Mapping
 
     #region Property
 
+    private static GitHubService GitHub => _gitHub ??= new();
+
     private static bool IsRunning => !_updateTask?.IsCompleted ?? false; // { private get; }
 
     public static MappingSettings Settings // { get; set; }
@@ -36,7 +39,7 @@ public static class Mapping
         get => _settings;
         set
         {
-            _settings = value ?? new();
+            _settings = value;
             _path = GetCombinedPath();
         }
     }
@@ -245,7 +248,7 @@ public static class Mapping
         {
             _updateTask = Task.Run(async () =>
             {
-                result = await DownloadAsync();
+                result = await GetJsonDownloadAsync();
                 if (result)
                 {
                     CreateMap();
@@ -268,7 +271,7 @@ public static class Mapping
 
         _updateTask = Task.Run(async () =>
         {
-            var result = await DownloadAsync();
+            var result = await GetJsonDownloadAsync();
             if (result)
             {
                 CreateMap();
@@ -277,13 +280,13 @@ public static class Mapping
     }
 
     /// <summary>
-    /// Downloads the lastet mapping file.
+    /// Downloads the lastet mapping file and persists it to a file.
     /// This method does not block the calling thread.
     /// </summary>
     /// <returns>Whether a newer version of the mapping file was successfully downloaded.</returns>
-    private static async Task<bool> DownloadAsync()
+    private static async Task<bool> GetJsonDownloadAsync()
     {
-        var content = await _httpClient.DownloadTextFileContentFromGitHubReleaseAsync(Properties.Resources.REPO_OWNER, Properties.Resources.REPO_NAME, Properties.Resources.RELEASE_ASSET);
+        var content = await GitHub.DownloadMappingJsonAsync();
         if (string.IsNullOrEmpty(content))
             return false;
 
@@ -297,10 +300,7 @@ public static class Mapping
             _ = File.WriteAllTextAsync(_path, content);
 #endif
         }
-        catch (IOException)
-        {
-            // Nothing to do.
-        }
+        catch (IOException) { } // Try again next time.
 
         _jsonDownload = MappingJson.Deserialize(content!);